requirements.txt
Flask==2.3.3
redis==5.0.1
supabase==1.2.0
requests==2.31.0
python-dotenv==1.0.0
selenium==4.15.0
psutil